adjusting Mocha type and create setupServer function
