+++ package.json
@@ -2,16 +2,19 @@
    "dependencies": {
       "@types/bcrypt": "5.0.0",
       "apollo-server": "3.10.1",
+      "axios": "0.27.2",
       "bcrypt": "5.0.1",
       "graphql": "16.6.0",
       "mocha": "10.0.0",
       "pg": "8.4.0",
-      "reflect-metadata": "0.1.13",
+      "reflect-metadata": "^0.1.13",
       "typeorm": "0.3.9",
       "typescript": "4.8.2"
    },
    "devDependencies": {
-      "@types/node": "16.11.10",
+      "@types/mocha": "^9.1.1",
+      "@types/node": "^16.11.10",
+      "mocha": "^10.0.0",
       "ts-node": "10.7.0",
       "ts-node-dev": "2.0.0",
       "typescript": "4.5.2"
@@ -19,7 +22,7 @@
    "scripts": {
       "start": "ts-node src/index.ts",
       "start:dev": "ts-node-dev --respawn --transpile-only src/index.ts",
-      "test":"ts-node test/index.ts",
+      "test": "mocha -r ts-node/register src/test/index.ts",
       "typeorm": "typeorm-ts-node-commonjs"
    }
 }
