+++ src/test/index.ts
@@ -1,0 +1,11 @@
+import { setupServer } from "../setup";
+
+before(async () => {
+  await setupServer();
+});
+
+describe("Query hello", () => {
+  it("should return hello word", async () => {
+    console.log("Iniciou o teste, testando...");
+  });
+});
