+++ tsconfig.json
@@ -11,5 +11,7 @@
       "emitDecoratorMetadata": true,
       "experimentalDecorators": true,
       "sourceMap": true
-   }
+   },
+   "exclude": ["node_modules"],
+   "include": [".eslintrc.js", "server.js", "src/**/*.ts"]
 }
